Deduplicated card XML parse code
The function to parse one-file-per-card data (parse_cardxml) was
already updated to parse the latest XML format, but the function to
parse all cards from a single file (parse_carddefs) was not. To avoid
problems like that in the future, I made both scenarios use the same
function (parse_cardentity).