Removed hearthy.bnet.* modules
This seems to be a partial server implementation using generated code
based on protocol buffer definitions. However, it also still depends
on the old protocol code (hearthy.protocol.mtypes a.o.) and the
generated code it expects is not the same as what I get as output
from the protoc compiler.

Since it is not trivial to migrate this code to use the new protocol
buffers generated code and I don't have an interest in server
functionality, I'm not going to put in the time required to properly
migrate and test the code. I'm removing the code rather than leaving
it in a broken state.

If you're looking for a Hearthstone account server, try Stove:
  https://github.com/HearthSim/Stove